Swift wrappers

+++ FaceWarpApp/triangulator.h
@@ -5,4 +5,4 @@
 //  Created by Thomas Nickson on 26/09/2015.
 //  Copyright © 2015 Phi Research. All rights reserved.
 //
-unsigned int * tidyIndices(int * edges, int nEdges, int * landMarks, int nFaces, int * nTris);
+int * tidyIndices(float * edges, int nEdges, float * landMarks, int nFaces, int * nTris);
